move utils, move matrix upload_to_s3 method to utils
