+++ flask_app/analyst/rank.py
@@ -6,7 +6,7 @@ from flask_app.analyst.base import _AnalystService
 
 from sppy.aws.aws_constants import PROJ_BUCKET
 from sppy.tools.s2n.spnet import SpNetAnalyses
-from sppy.tools.s2n.utils import (combine_errinfo, get_traceback)
+from sppy.tools.util.utils import (combine_errinfo, get_traceback)
 
 
 # .............................................................................
