+++ flask_app/broker/badge.py
@@ -6,7 +6,7 @@ from werkzeug.exceptions import (BadRequest, InternalServerError)
 from flask_app.broker.base import _BrokerService
 from flask_app.broker.constants import (ICON_CONTENT, ICON_DIR)
 from flask_app.common.s2n_type import APIService, S2nKey, ServiceProvider
-from sppy.tools.s2n.utils import get_traceback
+from sppy.tools.util.utils import get_traceback
 
 
 # .............................................................................
