+++ flask_app/broker/name.py
@@ -8,7 +8,7 @@ from flask_app.common.s2n_type import (
 from sppy.tools.provider.gbif import GbifAPI
 from sppy.tools.provider.itis import ItisAPI
 from sppy.tools.provider.worms import WormsAPI
-from sppy.tools.s2n.utils import get_traceback
+from sppy.tools.util.utils import get_traceback
 
 
 # .............................................................................
