+++ flask_app/common/base.py
@@ -2,7 +2,7 @@
 from flask import Flask
 from werkzeug.exceptions import BadRequest, InternalServerError
 
-from sppy.tools.s2n.utils import add_errinfo
+from sppy.tools.util.utils import add_errinfo
 from flask_app.common.s2n_type import APIEndpoint, APIService
 
 app = Flask(__name__)
