+++ sppy/tools/provider/idigbio.py
@@ -9,7 +9,7 @@ from flask_app.common.constants import ENCODING
 from sppy.tools.util.logtools import logit
 from sppy.tools.util.fileop import ready_filename
 from sppy.tools.provider.api import APIQuery
-from sppy.tools.s2n.utils import add_errinfo
+from sppy.tools.util.utils import add_errinfo
 
 
 # .............................................................................
