+++ sppy/tools/provider/mopho.py
@@ -4,7 +4,7 @@ from flask_app.common.s2n_type import (
     APIEndpoint, BrokerSchema, S2nKey, ServiceProvider)
 
 from sppy.tools.provider.api import APIQuery
-from sppy.tools.s2n.utils import add_errinfo
+from sppy.tools.util.utils import add_errinfo
 
 
 # .............................................................................
