+++ sppy/tools/provider/worms.py
@@ -8,7 +8,7 @@ from flask_app.common.s2n_type import (
 from flask_app.common.constants import URL_ESCAPES, ENCODING
 
 from sppy.tools.provider.api import APIQuery
-from sppy.tools.s2n.utils import add_errinfo
+from sppy.tools.util.utils import add_errinfo
 
 
 # .............................................................................
